Added features to frame by frame animation, added Timer, and fixed deltaMillis bug

+++ src/main/java/demo/Main.kt
@@ -24,7 +24,7 @@ fun main(args: Array<String>) {
 
 class StartState : State() {
     var position = -8.0
-    override fun update(delta: Float) {
+    override fun update(delta: Double) {
         position
         Game.viewHeight = Game.viewHeight + delta
     }
