+++ src/main/java/state/EmptyState.kt
@@ -3,6 +3,6 @@ package state
 import renderer.Painter
 
 object EmptyState : State(){
-    override fun update(delta: Float) { }
+    override fun update(delta: Double) { }
     override fun render(g: Painter) { }
 }
